requirements.txt
asgiref==3.6.0
async-timeout==4.0.2
attrs==22.2.0
autobahn==23.1.2
Automat==22.10.0
autopep8==2.0.2
backports.zoneinfo==0.2.1
cffi==1.15.1
channels==4.0.0
channels-redis==4.0.0
click==8.1.3
constantly==15.1.0
cryptography==40.0.1
daphne==4.0.0
Django==4.1.7
django-cors-headers==3.14.0
djangorestframework==3.14.0
gunicorn==20.1.0
h11==0.14.0
hyperlink==21.0.0
idna==3.4
incremental==22.10.0
msgpack==1.0.5
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.10.0
pycparser==2.21
pyOpenSSL==23.1.0
pytz==2022.7.1
redis==4.5.3
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
Twisted==22.10.0
txaio==23.1.1
typing-extensions==4.5.0
uvicorn==0.21.1
whitenoise==6.4.0
zope.interface==6.0
